First commit of the generic_gnupg branch.  This is taken from the old SVN repository.
The goal of the generic_gnupg branch is to modify gnupg.py into a form where it will be useful in many projects, not just the Nymserver.  This involves stripping out all the behaviour that's specific to Nymserver.  So far, the module is generic but nymserver.py is completely incompatible with it.